+++ SpetralAmlysisForms/y(n).cs
@@ -8,9 +8,9 @@ namespace SpetralAmlysisForms
 {
     class y_n_
     {
-        public float y(int n)
+        public double y(int n)
         {
-            float SUM=0;
+            double SUM=0;
             for (int i = 0; i < n; i++)
             {
                 SUM += 1;
